requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
python-dotenv==1.0.1
pydantic==2.9.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
scikit-learn==1.8.0
pandas==2.2.3
requests==2.32.3
aiofiles==24.1.0
bcrypt==4.0.1
